fix stockx_browse & complete scrape data

+++ stockx_browse.py
@@ -1,4 +1,5 @@
 import re
+import os
 import pandas as pd
 import httpx
 import asyncio
@@ -26,30 +27,37 @@ def get_product_browse(response):
 async def fetch_url(url):
     try:
         response = await client.get(url)
-        if response.status_code != 200:
+        if response.status_code == 200:
+            print(f"URL {url} succeeded")
+            return response.text, response.status_code
+        else:
             print(f"Skipping URL {url}: Status code {response.status_code}")
-            pass
-        print(f"URL {url} succeeded")
-        return response.text
+            return None, response.status_code
     except httpx.RequestError as e:
         print(f"An error occurred while requesting {url}: {e}")
 
-async def main(url):
-    response = await fetch_url(url)
-    name_list, link_list = get_product_browse(response)
-    return name_list, link_list
+async def main():
+    product_list = []
+    start_page = 1
+    end_page = 25
+    for page in range(start_page, end_page+1):
+        url = f"https://stockx.com/category/sneakers?page={page}"
+        response_text, status_code = await fetch_url(url)
+        if status_code != 200:
+            continue
+        name_list, link_list = get_product_browse(response_text)
+        
+        for nl, ll in zip(name_list, link_list):
+            product_list.append((nl, ll))  
 
-# browse 25 pages
-product_list = []
-for page in range(10, 26):
-    url = f"https://stockx.com/category/sneakers?page={page}"
-    name_list, link_list = asyncio.run(main(url))
-    
-    for nl, ll in zip(name_list, link_list):
-        product_list.append((nl, ll))  
+    product_browse_df = pd.DataFrame(product_list, columns=["product_name", "product_link"])
+    if os.path.exists('data/stockx_browse.csv'):
+        product_browse_df_read = pd.read_csv('data/stockx_browse.csv')
+        product_browse_df_merge = pd.concat([product_browse_df_read, product_browse_df])
+        product_browse_df_merge.drop_duplicates(inplace=True)
+        product_browse_df_merge.to_csv('data/stockx_browse.csv', index=False)
+    else:
+        product_browse_df.to_csv('data/stockx_browse.csv', index=False)
 
-product_browse_df = pd.DataFrame(product_list, columns=["product_name", "product_link"])
-product_browse_df_read = pd.read_csv('data/stockx_browse.csv')
-product_browse_df_merge = pd.concat([product_browse_df_read, product_browse_df])
-product_browse_df_merge.drop_duplicates(inplace=True)
-product_browse_df_merge.to_csv('data/stockx_browse.csv', index=False)
+if __name__ == '__main__':
+    asyncio.run(main())
